Added MPICH2 to Northpark.
Got all serial/merge 32/64 bit compilations working.

git-svn-id: svn://136.177.114.72/svn_GW/phreeqc/trunk/src@3490 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ basic.c
@@ -3274,7 +3274,7 @@ factor(struct LOC_exec * LINK)
 
 	case toklen:
 		s = strfactor(LINK);
-		n.UU.val = strlen(s);
+		n.UU.val = (double) strlen(s);
 		free(s);
 		break;
